+++ src/utils/d_array.c
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-void u_d_arr_ptr_init(d_array_ptr_t *arr, int64_t init_capacity) {
+void u_d_arr_ptr_init(d_array_ptr_t *arr, int32_t init_capacity) {
   arr->capacity = init_capacity;
   arr->length = 0;
   arr->arr = malloc(sizeof(void *) * init_capacity);
@@ -13,8 +13,8 @@ void u_d_arr_ptr_init(d_array_ptr_t *arr, int64_t init_capacity) {
   memset(arr->arr, 0, init_capacity);
 }
 
-void u_d_arr_ptr_add(d_array_ptr_t *arr, void *ptr, int64_t idx) {
-  int64_t new_capacity = idx * 2;
+void u_d_arr_ptr_add(d_array_ptr_t *arr, void *ptr, int32_t idx) {
+  int32_t new_capacity = idx * 2;
   if (idx > arr->capacity - 1) {
     arr->arr = realloc(arr->arr, new_capacity);
     memset(arr->arr+arr->capacity, 0, arr->capacity - new_capacity);
